include Source.c and fix compile bugs
