+++ cmake-build-debug/CMakeFiles/entranceSystem.dir/C.includecache
@@ -6,11 +6,19 @@
 
 #IncludeRegexTransform: 
 
+/Users/odedarlevski/Documents/GitHub/entranceSystem/Source.c
+stdio.h
+-
+time.h
+-
+
 /Users/odedarlevski/Documents/GitHub/entranceSystem/main.c
 stdio.h
 -
 stdlib.h
-/Users/odedarlevski/Documents/GitHub/entranceSystem/stdlib.h
+-
 string.h
-/Users/odedarlevski/Documents/GitHub/entranceSystem/string.h
+-
+Source.c
+/Users/odedarlevski/Documents/GitHub/entranceSystem/Source.c
