+++ cmake-build-debug/CMakeFiles/entranceSystem.dir/depend.internal
@@ -2,4 +2,5 @@
 # Generated by "Unix Makefiles" Generator, CMake Version 3.13
 
 CMakeFiles/entranceSystem.dir/main.c.o
+ /Users/odedarlevski/Documents/GitHub/entranceSystem/Source.c
  /Users/odedarlevski/Documents/GitHub/entranceSystem/main.c
