+++ cmake-build-debug/CMakeFiles/entranceSystem.dir/depend.make
@@ -1,5 +1,6 @@
 # CMAKE generated file: DO NOT EDIT!
 # Generated by "Unix Makefiles" Generator, CMake Version 3.13
 
+CMakeFiles/entranceSystem.dir/main.c.o: ../Source.c
 CMakeFiles/entranceSystem.dir/main.c.o: ../main.c
